backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.23
python-multipart==0.0.6
google-generativeai==0.3.2
googletrans==4.0.0rc1
python-dotenv==1.0.0
google-generativeai==0.3.2
googletrans==4.0.0rc1
requests==2.31.0
python-dotenv==1.0.0
